requirements.txt
ConfigArgParse==1.2.3
GitPython==3.1.26
joblib==0.17.0
lifelines==0.25.7
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
PyYAML==6.0
scikit_learn==1.0.2
seaborn==0.11.0
torch==1.8.0
